Finished resource binding backend for Vulkan. However, there is a bug with allocating descriptor sets
Former-commit-id: e202b13a14f961267fb5d15868788569ca2ab1ce [formerly f29081d724ebd602a93d381c79d89c1bd507783e]
Former-commit-id: 9855b87eee1e75b061aa61a5752693e54cbbfcd4

+++ fancy_core/CommandListDX12.cpp
@@ -1311,7 +1311,7 @@ namespace Fancy {
         for (uint iTableEntry = 0u; iTableEntry < descTable.myDescriptors.size(); ++iTableEntry)
           ASSERT(descTable.myDescriptors[iTableEntry].myCpuHandle.ptr != UINT_MAX, "Missing descriptor binding %d in descriptor table %d", iTableEntry, i);
 #endif  // FANCY_RENDERER_DEBUG
-        const DescriptorDX12 firstGpuVisibleDescriptor = CopyDescriptorsToDynamicHeapRange(descTable.myDescriptors.data(), descTable.myDescriptors.size());
+        const DescriptorDX12 firstGpuVisibleDescriptor = CopyDescriptorsToDynamicHeapRange(descTable.myDescriptors.data(), static_cast<uint>(descTable.myDescriptors.size()));
         if (myCurrentContext == CommandListType::Graphics)
           myCommandList->SetGraphicsRootDescriptorTable(i, firstGpuVisibleDescriptor.myGpuHandle);
         else
